+++ nope/lib/models/Media.php
@@ -5,13 +5,14 @@ namespace Nope;
 use RedBeanPHP\R as R;
 use Respect\Validation\Validator as v;
 use Respect\Validation\Exceptions\NestedValidationException;
+use Intervention\Image\ImageManagerStatic as Image;
 
 class Media extends Content {
 
   const MODELTYPE = 'media';
 
   function getPath() {
-    return NOPE_UPLOADS_DIR . $this->model->filename;
+    return NOPE_UPLOADS_DIR . $this->filename;
   }
 
   function getUrl($cache=true) {
@@ -23,6 +24,17 @@ class Media extends Content {
     $needle = 'image/';
     $length = strlen($needle);
     return (substr($this->mimetype, 0, $length) === $needle);
+  }
+
+  function getMetadata() {
+    if($this->isImage()) {
+      $path = $this->getPath();
+      return (object) [
+        'exif' => Image::make($path)->exif(),
+        'iptc' => Image::make($path)->iptc()
+      ];
+    }
+    return null;
   }
 
   function delete() {
@@ -50,6 +62,11 @@ class Media extends Content {
     }
   }
 
+  function beforeSave() {
+    $this->metadata = json_encode($this->getMetadata());
+    parent::beforeSave();
+  }
+
   function jsonSerialize() {
     $obj = parent::jsonSerialize();
     $obj->url = $this->getUrl();
@@ -58,6 +75,7 @@ class Media extends Content {
       $obj->preview->$key = $this->getPreview($key);
     }
     $obj->isImage = $this->isImage();
+    $obj->metadata = json_decode($this->metadata);
     unset($obj->cover);
     return $obj;
   }
@@ -79,10 +97,6 @@ class Media extends Content {
   static public function findAll($filters=null, $limit=-1, $offset=0, &$count=0, $orderBy='id desc') {
     $filters = (object) $filters;
     $params = [];
-    /*if($filters->role) {
-      $sql[] = 'role = ?';
-      $params[] = $filters->role;
-    }*/
     if($orderBy) {
       $sql[] = 'order by '.$orderBy;
     }